extract SkillWithExperiences
